+++ src/soho/dropdown/soho-dropdown.d.ts
@@ -20,14 +20,31 @@ interface SohoDropDownOptions {
   cssClass?: string;
 
   /**
+   * Search mode to use between 'startsWith' and 'contains', false will not allow client side filter
+   */
+  filterMode?: false | 'startsWith' | 'contains';
+
+  /**
+   * If true, displays <optgroup> headers in the list even if no selectable options are present underneath.
+   */
+  showEmptyGroupHeaders?: boolean;
+
+  /**
    * If in multiple mode, this setting sets a limit on the number of items that can be selected.
    */
   maxSelected?: number;
 
   /**
    * When the menu is opened, displays all selected options at the top of the list
+   *
+   * @deprectaed - use moveSelected.
    */
   moveSelectedToTop?: boolean;
+
+  /**
+   * Repl
+   */
+  moveSelected?: 'group' | 'all' | 'none';
 
   /**
    * If true, turns the dropdown into a multiple selection box; otherwise
@@ -48,6 +65,18 @@ interface SohoDropDownOptions {
   source?: SohoDropDownSourceFunction | Object | string;
 
   /**
+   * If a source method is defined, this flexible object can be passed into
+   * the source method, and augmented with parameters specific to the implementation.
+   */
+  sourceArguments?: any;
+
+  /**
+   * If set to true, will always perform an ajax call whenever the list is opened.  If false,
+   * the first AJAX call's results are cached.
+   */
+  reloadSourceOnOpen?: boolean;
+
+  /**
    * Allow an empty value to be selected, representing no selection.
    */
   empty?: boolean;
@@ -56,6 +85,12 @@ interface SohoDropDownOptions {
    * Typing Buffer Delay
    */
   delay?: number;
+
+  /**
+   * If set the width of the dropdown is limited to this pixel width.
+   * Fx 300 for the 300 px size fields. Default is size of the largest data.
+   */
+  maxWidth?: number;
 
 }